Refactor tests into alg package.
