+++ src/test/java/com/graphhopper/sna/alg/DFSTest.java
@@ -22,7 +22,7 @@
  * You should have received a copy of the GNU General Public License along with
  * GraphHopper-SNA. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.graphhopper.sna.analyzers;
+package com.graphhopper.sna.alg;
 
 import com.graphhopper.storage.Graph;
 import com.graphhopper.storage.GraphStorage;
